# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenMapX API",
    "description": "Generated from the API's own route table and the integration route registry. Do not edit by hand — run `pnpm openapi:generate`.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/integrations/overlay-satellite/tiles/{layerId}/{date}/{z}/{y}/{x}": {
      "get": {
        "operationId": "getApiIntegrationsOverlaySatelliteTilesLayerIdDateZYX",
        "parameters": [
          {
            "name": "layerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /tiles/:layerId/:date/:z/:y/:x (overlay-satellite)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "overlay-satellite",
        "x-openmapx-source": "integrations/overlay-satellite/index.ts"
      }
    },
...
    "/api/integrations/overlay-weather/radar/tile/{z}/{x}/{y}": {
      "get": {
        "operationId": "getApiIntegrationsOverlayWeatherRadarTileZXY",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /radar/tile/:z/:x/:y (overlay-weather)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "overlay-weather",
        "x-openmapx-source": "integrations/overlay-weather/index.ts"
      }
    },
...
    "/api/integrations/street-level-imagery-mapillary/tiles/{z}/{x}/{y}": {
      "get": {
        "operationId": "getApiIntegrationsStreetLevelImageryMapillaryTilesZXY",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /tiles/:z/:x/:y (street-level-imagery-mapillary)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "street-level-imagery-mapillary",
        "x-openmapx-source": "integrations/street-level-imagery-mapillary/index.ts"
      }
    },
...
    "/api/integrations/street-level-imagery-panoramax/tiles/{z}/{x}/{y}": {
      "get": {
        "operationId": "getApiIntegrationsStreetLevelImageryPanoramaxTilesZXY",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /tiles/:z/:x/:y (street-level-imagery-panoramax)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "street-level-imagery-panoramax",
        "x-openmapx-source": "integrations/street-level-imagery-panoramax/index.ts"
      }
    },
...
    "/api/integrations/transit/routes/{id}": {
      "get": {
        "operationId": "getApiIntegrationsTransitRoutesId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /routes/:id (transit)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "transit",
        "x-openmapx-source": "integrations/transit/index.ts"
      }
    },
...
    "/api/tiles/cycling-routes/{z}/{x}/{y}.png": {
      "get": {
        "operationId": "getApiTilesCyclingRoutesZXYPng",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{1,2}$"
            },
            "in": "path",
            "name": "z",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "in": "path",
            "name": "x",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "in": "path",
            "name": "y",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "tiles"
        ],
        "x-openmapx-auth": "public"
      }
    },
    "/api/tiles/cyclosm/{z}/{x}/{y}.png": {
      "get": {
        "operationId": "getApiTilesCyclosmZXYPng",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{1,2}$"
            },
            "in": "path",
            "name": "z",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "in": "path",
            "name": "x",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "in": "path",
            "name": "y",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "tiles"
        ],
        "x-openmapx-auth": "public"
      }
    },
    "/api/tiles/terrain/{z}/{x}/{y}.png": {
      "get": {
        "operationId": "getApiTilesTerrainZXYPng",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{1,2}$"
            },
            "in": "path",
            "name": "z",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "in": "path",
            "name": "x",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "in": "path",
            "name": "y",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "tiles"
        ],
        "x-openmapx-auth": "public"
      }
    },
...
  }
}